[TASK] 1. Integrated the following functionalities in the error signalling module: 		- status entry updation 		- emergency queue entry updation 		- static error bit field updation        2. Added events and handlers for 1001 and 1003 entry updation in OBD and corresponding error signalling

+++ examples/demo_cn_console/app.c
@@ -75,13 +75,13 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 /* structure for input process image */
 typedef struct
 {
-   BYTE    digitalIn;
+   UINT    digitalIn;
 } PI_IN;
 
 /* structure for output process image */
 typedef struct
 {
-   BYTE    digitalOut;
+   UINT    digitalOut;
 } PI_OUT;
 
 //------------------------------------------------------------------------------
@@ -92,8 +92,8 @@ static PI_IN*   pProcessImageIn_l;
 static PI_OUT*  pProcessImageOut_l;
 
 /* application variables */
-static BYTE    digitalIn_g;                 // 8 bit digital input
-static BYTE    digitalOut_g;                // 8 bit digital output
+static UINT    digitalIn_g;                 // 8 bit digital input
+static UINT    digitalOut_g;                // 8 bit digital output
 
 //------------------------------------------------------------------------------
 // local function prototypes
